Add the new funcionality about the favorites pokemon and add action for the bottom buttons into ListView

+++ Pokemon/Model/PokemonData.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 struct PokemonData: Codable {
-    let weight: Float
-    let height: Float
+    let weight: Double
+    let height: Double
     let types: [PokemonType]?
     let sprites: Sprite?
 }
